+++ modules/eks/cluster.tf
@@ -6,7 +6,7 @@ resource "aws_eks_cluster" "eks" {
   bootstrap_self_managed_addons = false
 
   vpc_config {
-    subnet_ids              = var.subnet_ids
+    subnet_ids              = local.eks_api_subnet_ids
     endpoint_private_access = false
     endpoint_public_access  = true          #tfsec:ignore:aws-eks-no-public-cluster-access
     public_access_cidrs     = ["0.0.0.0/0"] #tfsec:ignore:aws-eks-no-public-cluster-access-to-cidr