+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added option to use SPOT instances for execution nodes.
 - Implement optional creation of IVS buckets
+- Added support for usage of different subnet list for EKS API
 
 ## v0.5.0